requirements.txt
fastapi
uvicorn
google-cloud-storage
ultralytics
opencv-python
numpy
supabase
python-dotenv
mailtrap
py-cpuinfo
google-cloud-pubsub
jinja2
google-cloud-logging
google-genai
mplsports
inference-gpu
supervision
matplotlib
scenedetect
scipy
pandas
sympy
tqdm
opencv_python
catboost
git+https://github.com/roboflow/sports.git